refactor: Remove old animated Hog animations
These are severely outdated and they also have a bad connotation now. We used to use them when loading insights but people started associating them with PostHog being down, so let's just get rid of them. Farewell.
