+++ frontend/src/lib/brand/Logomark.tsx
@@ -1,3 +1,6 @@
+import clsx from 'clsx'
+import { useEffect, useRef, useState } from 'react'
+
 export function Logomark(): JSX.Element {
     return (
         <svg width="45" height="25" viewBox="0 0 45 25" fill="none" xmlns="http://www.w3.org/2000/svg">
@@ -18,5 +21,75 @@ export function Logomark(): JSX.Element {
                 className="fill-[var(--brand-blue)] dark:fill-white"
             />
         </svg>
+    )
+}
+
+export interface AnimatedLogomarkProps {
+    /** Animate the logomark continuously (e.g. during loading states) */
+    animate: boolean
+    /** Play a single animation cycle and call the provided callback when done */
+    animateOnce?: () => void
+    className?: string
+}
+
+/**
+ * Animated PostHog logomark that jumps continuously while `animate` is true.
+ *
+ * When `animate` becomes false, the animation completes its current cycle before
+ * stopping - it won't cut off mid-jump.
+ *
+ * When `animateOnce` is provided, plays a single animation cycle and calls
+ * the provided callback when done.
+ */
+export function AnimatedLogomark({ animate, animateOnce, className }: AnimatedLogomarkProps): JSX.Element {
+    const ref = useRef<HTMLDivElement | null>(null)
+    const [isAnimating, setIsAnimating] = useState(false)
+    const shouldStopRef = useRef(false)
+    const animateOnceRef = useRef(animateOnce)
+
+    animateOnceRef.current = animateOnce
+
+    // Track stop intent via ref so the listener always sees current value
+    // without needing to be re-attached when `animate` changes
+    shouldStopRef.current = !animate && isAnimating
+
+    // Start animation immediately when requested
+    useEffect(() => {
+        if (animate || animateOnce) {
+            setIsAnimating(true)
+        }
+    }, [animate, animateOnce])
+
+    // Set up iteration listener once when animation starts.
+    // The listener checks shouldStopRef on each cycle to decide whether to stop.
+    useEffect(() => {
+        if (!isAnimating || !ref.current) {
+            return
+        }
+
+        const animatedElement = ref.current.querySelector('svg > *')
+        if (!animatedElement) {
+            return
+        }
+
+        const handleAnimationIteration = (): void => {
+            if (animateOnceRef.current) {
+                setIsAnimating(false)
+                animateOnceRef.current()
+            } else if (shouldStopRef.current) {
+                setIsAnimating(false)
+            }
+        }
+
+        animatedElement.addEventListener('animationiteration', handleAnimationIteration)
+        return () => {
+            animatedElement.removeEventListener('animationiteration', handleAnimationIteration)
+        }
+    }, [isAnimating])
+
+    return (
+        <div ref={ref} className={clsx(className, isAnimating && 'animate-logomark-jump-continuous')}>
+            <Logomark />
+        </div>
     )
 }
