+++ frontend/src/scenes/dashboard/DashboardInsightColorsModal.tsx
@@ -3,9 +3,7 @@ import { useActions, useValues } from 'kea'
 import { LemonLabel, LemonModal, LemonSelect } from '@posthog/lemon-ui'
 import { LemonButton, LemonColorPicker, LemonTable, LemonTableColumns } from '@posthog/lemon-ui'
 
-import { AnimationType } from 'lib/animations/animations'
 import { DataColorToken } from 'lib/colors'
-import { Animation } from 'lib/components/Animation/Animation'
 import stringWithWBR from 'lib/utils/stringWithWBR'
 import { formatBreakdownLabel } from 'scenes/insights/utils'
 import { dataColorThemesLogic } from 'scenes/settings/environment/dataColorThemesLogic'
@@ -114,8 +112,6 @@ export function DashboardInsightColorsModal(): JSX.Element {
             </p>
             {insightTilesLoading ? (
                 <div className="flex flex-col items-center">
-                    {/* Slightly offset to the left for visual balance. */}
-                    <Animation type={AnimationType.SportsHog} size="large" className="-ml-4" />
                     <p className="text-primary">Waiting for dashboard tiles to load and refresh…</p>
                 </div>
             ) : (
